Add stype to types: div, p, and button
In this lesson I had to create a new style sheet and style types
following the instructions. The style for each type were:
- div: a red background, white text, a font size of 32px, center
aligned, and bold
- p: a green background, white text, and a font size of 18px
- button: an orange background and a font size of 18px

The div type was styled using the external method, that is, using the
styles.css file.
To style the paragraph (p), I used the element style inside the HTML
file.
And the button was styled using the inline method.